menu context provided. But learned swiper displays renders all screens at same time. will need to do another refactor
